Tests/LibWeb/Text/expected/css/transition-ordering.txt

Order: bottom, height, left, right, top, width
